% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-gldrm.R
\name{print.gldrm}
\alias{print.gldrm}
\title{Print summary of gldrm fit}
\usage{
\method{print}{gldrm}(x, digits = 3, ...)
}
\arguments{
\item{x}{S3 object of class "gldrm", returned from the \code{gldrm} function.}

\item{digits}{Number of digits for rounding.}

\item{...}{Unused. Additional arguments for print method.}
}
\description{
Prints fitted coefficients and standard errors, along with a likelihood ratio
test against the null model.
}
