% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldrmCI.R
\name{print.gldrmCI}
\alias{print.gldrmCI}
\title{Print confidence interval}
\usage{
\method{print}{gldrmCI}(x, digits = 3, ...)
}
\arguments{
\item{x}{An S3 object of class 'gldrmCI'.}

\item{digits}{Number of digits for rounding.}

\item{...}{Not used. Additional arguments for print method.}
}
\description{
Print method for gldrmCI objects.
}
