% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-gldrmFit.R
\name{print.gldrmFit}
\alias{print.gldrmFit}
\title{Print summary of gldrm fit}
\usage{
\method{print}{gldrmFit}(x, digits = 3, ...)
}
\arguments{
\item{x}{S3 object of class gldrmFit, returned from the gldrm function.}

\item{digits}{Number of digits for rounding.}

\item{...}{Unused. Additional arguments for print method.}
}
\description{
Prints fitted coefficients and standard errors, along with a likelihood ratio
test against the null model.
}
