% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldrmLRT.R
\name{print.gldrmLRT}
\alias{print.gldrmLRT}
\title{Print likelihood ratio test results}
\usage{
\method{print}{gldrmLRT}(x, digits = 3, ...)
}
\arguments{
\item{x}{S3 object of class 'gldrmLRT', returned from the \code{gldrmLRT} function.}

\item{digits}{Number of digits for rounding.}

\item{...}{Not used. Additional arguments for print method.}
}
\description{
Print method for gldrmLRT objects. Prints results of a likelihood ratio F-test
between nested models.
}
